requirements.txt
absl-py==1.4.0
altair==5.1.1
annotated-types==0.5.0
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.2
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
confection==0.1.1
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
idna==3.4
imageio==2.31.2
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.13.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
lottie==0.7.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
pathy==0.10.2
Pillow==9.5.0
preshed==3.0.8
progressbar2==4.2.0
protobuf==4.24.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-utils==3.7.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.10.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.15.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
streamlit==1.26.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
tensorlayerx==0.5.8
termcolor==2.3.0
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.8.30
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.21.2
wasabi==1.1.2
Werkzeug==2.3.7
wrapt==1.15.0
zipp==3.16.2
